% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almshouse_admissions.R
\docType{data}
\name{almshouse_admissions}
\alias{almshouse_admissions}
\title{Philadelphia Almshouse Admissions 1796-1803}
\format{
One data frame with 201 variables and 3886 observations.
}
\usage{
almshouse_admissions
}
\description{
A sample of people admitted into the Philadelphia Almshouse 1796-1803.
This contains a this is a sample from the Index, including last names that
started A through G as well as a handful of people indicated by race. This
data was created from and Index of Admissions to the Philadelphia Almshouse.
The original record is in Philadelphia City Archives. We used a microfilm
of that record to create the dataset. See \href{https://repository.upenn.edu/mead/42/}
{the original repository} for the dataset's codebook.
}
\examples{
data(judges)
head(judges)
}
\references{
This data is taken from the
\href{https://repository.upenn.edu/mead/42/}{Almshouse Admissions 1796-1803
repository at the Magazine of Early American Datasets}.
}
\author{
Billy G. Smith
}
\keyword{datasets}
