% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BostonWomenVoters.R
\docType{data}
\name{BostonWomenVoters}
\alias{BostonWomenVoters}
\title{1920 Boston Women's Voter Register
This data is derived from the Boston Archives and Records Management}
\format{
One data frame with 43 variables and 10702 observations.
}
\usage{
BostonWomenVoters
}
\description{
The Women's Voter Register Dataset is created from Election Department
registers used to register women voters in 1920 after the passage of the 19th
Amendment. The dataset contains information about newly registered women
voters including name, address, place of birth, occupation, place of work,
naturalization information, and closest male relative. This dataset is in
progress and is updated periodically as additional voter registers are
transcribed. The data set included in this package was downloaded in December
2022.
}
\examples{
data("BostonWomenVoters")
}
\references{
This data is taken from the
\href{https://data.boston.gov/dataset/1920-women-s-voter-register}{Boston
Archives and Records Management}.
}
\keyword{datasets}
