% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charlestonManumissions.R
\docType{data}
\name{charlestonManumissions}
\alias{charlestonManumissions}
\title{Manumissions Data, Charleston District, South Carolina, 1776-1800}
\format{
One data frame with 201 variables and 3886 observations.
}
\usage{
charlestonManumissions
}
\description{
This database contains all of the manumissions filed in Charleston District
from 1776-1800 contained in the Miscellaneous Records section of the South
Carolina Department of Archives and History. After 1800, South Carolina
changed its manumission laws to require court approval for manumission,
causing the paperwork associated with them to be filed elsewhere likely
with the no longer extent records of the court of magistrates and
freeholders. See \href{https://repository.upenn.edu/mead/34/}
{the original repository} for more info.
}
\examples{
data("CharlestonManumissions")
head(CharlestonManumissions)
                                                                                                                                                                                                                                                                                                                                                             For more information, see John Garrison Marks, "Race and Freedom in the African Americas: Free People of Color and Social Mobility in Cartagena and Charleston," PhD Dissertation (2016, Rice University).
}
\references{
This data is taken from the
\href{https://repository.upenn.edu/mead/42/}{Almshouse Admissions 1796-1803
repository at the Magazine of Early American Datasets}.
}
\author{
Billy G. Smith
}
\keyword{datasets}
