% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalMethodsData-package.R
\docType{package}
\name{DigitalMethodsData}
\alias{DigitalMethodsData}
\title{Datasets for Digital Methods II at Clemson University}
\description{
This package provides sample datasets for the use of students in History 8510
at Clemson University. It is similar to the \pkg{datasets} and the
\pkg{histdata} package.
}
\section{References}{
 References for each dataset included here is included on
the documentation page.
}

