% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gayguides.R
\docType{data}
\name{gayguides}
\alias{gayguides}
\title{Bob Damron's Address Books}
\format{
One data frame with 14 variables and 60698 observations.
}
\usage{
gayguides
}
\description{
This dataset contains information transcribed from Bob Damron's Address
Books from 1965 through 1985. Last updated December 2022.
}
\examples{
data("gayguides")
head(gayguides)
}
\keyword{datasets}
