% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judges.R
\docType{data}
\name{judges}
\alias{judges}
\title{Federal judges in the United States of America}
\format{
One data frame with 201 variables and 3886 observations.
}
\usage{
judges
}
\description{
This dataset contains information about the appointments and careers of all
federal judges in United States history since 1789. It includes judges who
"judges presidentially appointed during good behavior who have served since
1789 on the U.S. District Courts, the U.S. Courts of Appeals, the Supreme
Court of the United States,  the former U.S. Circuit Courts, and the federal
judiciary's courts of special jurisdiction." This dataset is left purposely
messy for History 8510.
}
\examples{
data("judges")
head(judges)
}
\references{
This data is taken from the
\href{https://www.fjc.gov/history/judges/biographical-directory-article-iii-federal-judges-export}{Biographical
Directory of Federal Judges, 1789-present}.
}
\keyword{datasets}
