% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepopulations.R
\docType{data}
\name{statepopulations}
\alias{statepopulations}
\title{Populations of US states and territories, 1790-2010}
\format{
An object of class \code{data.frame} with 84 rows and 28 columns.
}
\usage{
statepopulations
}
\description{
Population figures for US states and territories from the decennial census.
}
\examples{
head(state_populations)
}
\references{
This dataset has been gathered by the
\href{https://www.nhgis.org/}{NHGIS}. Minnesota Population Center,
\emph{National Historical Geographic Information System: Version 2.0}
(Minneapolis: University of Minnesota, 2011).
}
\keyword{datasets}
