% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UndergroundRR.R
\docType{data}
\name{undergroundRR}
\alias{undergroundRR}
\title{Underground RR Data
The names/data of almost 1,000 freedom seekers who sought refuge with
abolitionist William Still in 1850s.}
\format{
A data frame with 997 observations and 18 variables.
}
\usage{
undergroundRR
}
\description{
Underground RR Data
The names/data of almost 1,000 freedom seekers who sought refuge with
abolitionist William Still in 1850s.
}
\examples{
head(undergroundRR)

}
\references{
This data was transcribed by Nick Sacco from the Appendix of
William C. Kashatus's \emph{William Still: The Underground Railroad and the Angel
at Philadelphia.} You can read more about this dataset on
\href{https://pastexplore.wordpress.com/2022/11/03/william-stills-underground-railroad-data-now-freely-available-online/}
{Sacco's blog}
}
\author{
\href{https://pastexplore.wordpress.com}{Nick Sacco}
}
\keyword{datasets}
